requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
